[DDW-677] Move app version parts from AppStore to NodeUpdateStore
