+++ source/renderer/app/actions/app-actions.js
@@ -11,5 +11,4 @@ export default class AppActions {
   getGpuStatus: Action<any> = new Action();
   initAppEnvironment: Action<any> = new Action();
   toggleBlockConsolidationStatusScreen: Action<any> = new Action();
-  getLatestAvailableAppVersion: Action<any> = new Action();
 }
