+++ source/renderer/app/actions/node-update-actions.js
@@ -7,4 +7,5 @@ export default class NodeUpdateActions {
   acceptNodeUpdate: Action<any> = new Action();
   postponeNodeUpdate: Action<any> = new Action();
   toggleNodeUpdateNotificationExpanded: Action<any> = new Action();
+  getLatestAvailableAppVersion: Action<any> = new Action();
 }
