+++ source/renderer/app/api/nodes/requests/getLatestAppVersionInfo.js
